+++ Holovibes/tools.cu
@@ -1,7 +1,6 @@
 #include "stdafx.h"
 #include "tools.cuh"
 
-
 // CONVERSION FUNCTIONS
 
 __global__ void image_2_complex8(cufftComplex* res, unsigned char* data, int size, float *sqrt_tab)
@@ -24,8 +23,8 @@ __global__ void image_2_complex16(cufftComplex* res, unsigned short* data, int s
 
   while (index < size)
   {
-    res[index].x = sqrt_tab[data[index]]; 
-    res[index].y = sqrt_tab[data[index]]; 
+    res[index].x = sqrt_tab[data[index]];
+    res[index].y = sqrt_tab[data[index]];
     index += blockDim.x * gridDim.x;
   }
 }
@@ -114,7 +113,7 @@ void shift_corners(unsigned short **input, int size_x, int size_y)
   cudaFree(output);
 }
 
-__global__ void kernel_endianness_conversion(unsigned short* input, unsigned short* output, unsigned int size)
+__global__ void kernel_endianness_conversion(unsigned short* input, unsigned short* output, size_t size)
 {
   unsigned int index = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -126,7 +125,7 @@ __global__ void kernel_endianness_conversion(unsigned short* input, unsigned sho
   }
 }
 
-__global__ void fft2_make_u_v(float pasu, float pasv, float *u, float *v, unsigned int size_x , unsigned int size_y)
+__global__ void fft2_make_u_v(float pasu, float pasv, float *u, float *v, unsigned int size_x, unsigned int size_y)
 {
   unsigned int index = blockIdx.x * blockDim.x + threadIdx.x;
   while (index < size_x || index < size_y)
@@ -140,7 +139,6 @@ __global__ void fft2_make_u_v(float pasu, float pasv, float *u, float *v, unsign
     else
       rounded = entire_part;
 
-
     if (index < size_x)
       u[index] = ((index - 1) - rounded) * pasu;
     if (index < size_y)
@@ -148,9 +146,9 @@ __global__ void fft2_make_u_v(float pasu, float pasv, float *u, float *v, unsign
     index += blockDim.x * gridDim.x;
   }
 }
-// output_u size_x * size_y 
-// output_v size_x * size_y
 
+// output_u size_x * size_y
+// output_v size_x * size_y
 __global__ void meshgrind_square(float *input_u, float *input_v, float *output_u, float *output_v, unsigned int size_x, unsigned int size_y)
 {
   unsigned int index = blockIdx.x * blockDim.x + threadIdx.x;
@@ -171,5 +169,5 @@ void endianness_conversion(unsigned short* input, unsigned short* output, unsign
   if (blocks > max_blocks)
     blocks = max_blocks - 1;
 
-  kernel_endianness_conversion << <blocks, threads >> >(input, output, size);
+  kernel_endianness_conversion <<<blocks, threads >>>(input, output, size);
 }
